Restructure API documentation: add lots of lists with links
