+++ kikuchipy/crystallography/__init__.py
@@ -16,6 +16,8 @@
 # You should have received a copy of the GNU General Public License
 # along with kikuchipy. If not, see <http://www.gnu.org/licenses/>.
 
+"""Crystallographic computations not in diffpy.structure."""
+
 from diffpy.structure import Lattice
 import numpy as np
