+++ kikuchipy/detectors/__init__.py
@@ -16,6 +16,8 @@
 # You should have received a copy of the GNU General Public License
 # along with kikuchipy. If not, see <http://www.gnu.org/licenses/>.
 
+"""An EBSD detector and related quantities."""
+
 from kikuchipy.detectors.ebsd_detector import EBSDDetector
 
 __all__ = ["EBSDDetector"]
