+++ kikuchipy/draw/__init__.py
@@ -16,6 +16,8 @@
 # You should have received a copy of the GNU General Public License
 # along with kikuchipy. If not, see <http://www.gnu.org/licenses/>.
 
+"""Create HyperSpy markers to add to signals."""
+
 from kikuchipy.draw import markers
 
 __all__ = ["markers"]
