+++ kikuchipy/filters/__init__.py
@@ -16,6 +16,8 @@
 # You should have received a copy of the GNU General Public License
 # along with kikuchipy. If not, see <http://www.gnu.org/licenses/>.
 
+"""Pattern filters used on signals."""
+
 from kikuchipy.filters.window import (
     distance_to_origin,
     modified_hann,
