+++ kikuchipy/generators/__init__.py
@@ -16,14 +16,14 @@
 # You should have received a copy of the GNU General Public License
 # along with kikuchipy. If not, see <http://www.gnu.org/licenses/>.
 
-from kikuchipy.generators.virtual_bse_generator import VirtualBSEGenerator
-from kikuchipy.generators import util
+"""Generate signals and simulations, sometimes *from* other signals."""
+
 from kikuchipy.generators.ebsd_simulation_generator import (
     EBSDSimulationGenerator,
 )
+from kikuchipy.generators.virtual_bse_generator import VirtualBSEGenerator
 
 __all__ = [
     "EBSDSimulationGenerator",
-    "util",
     "VirtualBSEGenerator",
 ]
