+++ kikuchipy/generators/virtual_bse_generator.py
@@ -29,10 +29,10 @@ import numpy as np
 
 from kikuchipy.signals import EBSD, LazyEBSD
 from kikuchipy.signals import VirtualBSEImage
-from kikuchipy.generators.util import (
-    get_rgb_image as get_rgb_image_from_arrays,
+from kikuchipy.generators._transfer_axes import (
     _transfer_navigation_axes_to_signal_axes,
 )
+from kikuchipy.pattern import rescale_intensity
 
 
 class VirtualBSEGenerator:
@@ -120,7 +120,7 @@ class VirtualBSEGenerator:
             Output data type, either np.uint16 or np.uint8 (default).
         kwargs
             Keyword arguments passed to
-            :func:` ~kikuchipy.generators.util.virtual_bse.get_rgb_image`.
+            :func:`~kikuchipy.generators.virtual_bse_generator.get_rgb_image`.
 
         Returns
         -------
@@ -131,7 +131,6 @@ class VirtualBSEGenerator:
         --------
         kikuchipy.signals.EBSD.plot_virtual_bse_intensity,
         kikuchipy.signals.EBSD.get_virtual_bse_intensity,
-        kikuchipy.generators.util.get_rgb_image
 
         Notes
         -----
@@ -160,7 +159,7 @@ class VirtualBSEGenerator:
         if alpha is not None and isinstance(alpha, Signal2D):
             alpha = alpha.data
 
-        rgb_image = get_rgb_image_from_arrays(
+        rgb_image = get_rgb_image(
             channels=channels,
             normalize=normalize,
             alpha=alpha,
@@ -341,3 +340,109 @@ class VirtualBSEGenerator:
         pattern.add_marker(markers, permanent=True)
 
         return pattern
+
+
+def normalize_image(
+    image: np.ndarray,
+    add_bright: int = 0,
+    contrast: int = 1.0,
+    dtype_out: Union[np.uint8, np.uint16] = np.uint8,
+) -> np.ndarray:
+    """Normalize an image's intensities to a mean of 0 and a standard
+    deviation of 1, with the possibility to also scale by a contrast
+    factor and shift the brightness values.
+
+    Clips intensities to uint8 data type range, [0, 255].
+
+    Adapted from the aloe/xcdskd package.
+
+    Parameters
+    ----------
+    image
+        Image to normalize.
+    add_bright
+        Brightness offset. Default is 0.
+    contrast
+        Contrast factor. Default is 1.0.
+    dtype_out
+        Output data type, either np.uint16 or np.uint8 (default).
+
+    Returns
+    -------
+    image_out : np.ndarray
+    """
+    dtype_max = np.iinfo(dtype_out).max
+
+    offset = (dtype_max // 2) + add_bright
+    contrast *= dtype_max * 0.3125
+    median = np.median(image)
+    std = np.std(image)
+    normalized_image = offset + ((contrast * (image - median)) / std)
+
+    return np.clip(normalized_image, 0, dtype_max)
+
+
+def get_rgb_image(
+    channels: List[np.ndarray],
+    percentiles: Optional[Tuple] = None,
+    normalize: bool = True,
+    alpha: Optional[np.ndarray] = None,
+    dtype_out: Union[np.uint8, np.uint16] = np.uint8,
+    **kwargs,
+) -> np.ndarray:
+    """Return an RGB image from three numpy arrays, with a potential
+    alpha channel.
+
+    Parameters
+    ----------
+    channels
+        A list of np.ndarray for the red, green and blue channel,
+        respectively.
+    normalize
+        Whether to normalize the individual `channels` before
+        RGB image creation.
+    alpha
+        Potential alpha channel. If None (default), no alpha channel
+        is added to the image.
+    percentiles
+        Whether to apply contrast stretching with a given percentile
+        tuple with percentages, e.g. (0.5, 99.5), after creating the
+        RGB image. If None (default), no contrast stretching is
+        performed.
+    dtype_out
+        Output data type, either np.uint16 or np.uint8 (default).
+    kwargs :
+        Keyword arguments passed to
+        :func:`~kikuchipy.generators.virtual_bse_generator.normalize_image`.
+
+    Returns
+    -------
+    rgb_image : np.ndarray
+        RGB image.
+    """
+    n_channels = 3
+    rgb_image = np.zeros(channels[0].shape + (n_channels,), np.float32)
+    for i, channel in enumerate(channels):
+        if normalize:
+            channel = normalize_image(
+                channel.astype(np.float32), dtype_out=dtype_out, **kwargs
+            )
+        rgb_image[..., i] = channel
+
+    # Apply alpha channel if desired
+    if alpha is not None:
+        alpha_min = np.nanmin(alpha)
+        rescaled_alpha = (alpha - alpha_min) / (np.nanmax(alpha) - alpha_min)
+        for i in range(n_channels):
+            rgb_image[..., i] *= rescaled_alpha
+
+    # Rescale to fit data type range
+    if percentiles is not None:
+        in_range = tuple(np.percentile(rgb_image, q=percentiles))
+    else:
+        in_range = None
+    rgb_image = rescale_intensity(
+        rgb_image, in_range=in_range, dtype_out=dtype_out
+    )
+
+    return rgb_image.astype(dtype_out)
