+++ kikuchipy/io/__init__.py
@@ -15,3 +15,5 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with kikuchipy. If not, see <http://www.gnu.org/licenses/>.
+
+"""Read and write signals from and to file."""
