+++ kikuchipy/io/plugins/emsoft_ebsd.py
@@ -16,6 +16,8 @@
 # You should have received a copy of the GNU General Public License
 # along with kikuchipy. If not, see <http://www.gnu.org/licenses/>.
 
+"""Read support for simulated EBSD patterns in EMsoft's HDF5 format."""
+
 from typing import List, Tuple, Union
 
 import dask.array as da
