+++ kikuchipy/io/plugins/h5ebsd.py
@@ -16,6 +16,8 @@
 # You should have received a copy of the GNU General Public License
 # along with kikuchipy. If not, see <http://www.gnu.org/licenses/>.
 
+"""Read/write support for EBSD patterns in some HDF5 file formats."""
+
 import os
 from typing import Union, List, Tuple, Optional, Dict
 import warnings
