+++ kikuchipy/io/plugins/nordif.py
@@ -16,6 +16,8 @@
 # You should have received a copy of the GNU General Public License
 # along with kikuchipy. If not, see <http://www.gnu.org/licenses/>.
 
+"""Read/write support for EBSD patterns in NORDIF's binary format."""
+
 import datetime
 import os
 import re
