+++ kikuchipy/pattern/correlate.py
@@ -16,6 +16,8 @@
 # You should have received a copy of the GNU General Public License
 # along with kikuchipy. If not, see <http://www.gnu.org/licenses/>.
 
+"""Compute similarities between EBSD patterns."""
+
 from typing import Union
 
 from dask.array import Array
