+++ kikuchipy/signals/__init__.py
@@ -16,6 +16,10 @@
 # You should have received a copy of the GNU General Public License
 # along with kikuchipy. If not, see <http://www.gnu.org/licenses/>.
 
+"""Experimental and simulated diffraction patterns and virtual
+backscatter electron images.
+"""
+
 from kikuchipy.signals import util
 from kikuchipy.signals.ebsd import EBSD, LazyEBSD
 from kikuchipy.signals.ebsd_master_pattern import (
